renaming and reorganizing set based algo files
